now uses an array instead and simplified even more.
this might be the last edit to the graph file but we'll see what happens. I will probably need to put the algorithm inside the graph class to reduce the need for any get classes. I am also working on a pretty cool way to get it to read in a graph. that will be in the main file.