Create main.java   Beta 1.00
this creates the graph off of a file. later I may make this so it runs off the command line and takes in the filename. the file that is used to make the graph is created by another java thing I found called GraphDraw. you can make a graph in the gui and then save it as a list which is then read by main. the file must be blaced in the project folder, not in any sub folders.